feat: Remove sdkinfo getter on client move everything to integration
