+++ packages/core/src/interfaces.ts
@@ -2,7 +2,6 @@ import { Scope } from '@sentry/hub';
 import {
   Breadcrumb,
   Integration,
-  SdkInfo,
   SentryEvent,
   SentryResponse,
   Transport,
@@ -217,9 +216,6 @@ export interface Client<O extends Options = Options> {
 
   /** Returns the current options. */
   getOptions(): O;
-
-  /** Returns the current used SDK version and name. */
-  getSdkInfo(): SdkInfo;
 }
 
 /**
