+++ packages/core/test/mocks/client.ts
@@ -1,12 +1,6 @@
-import { SdkInfo } from '@sentry/types';
 import { BaseClient } from '../../src/base';
 import { initAndBind } from '../../src/sdk';
 import { TestBackend, TestOptions } from './backend';
-
-export const TEST_SDK = {
-  name: 'sentry-test',
-  version: '0.0.0-dev',
-};
 
 export class TestClient extends BaseClient<TestBackend, TestOptions> {
   public static instance?: TestClient;
@@ -14,10 +8,6 @@ export class TestClient extends BaseClient<TestBackend, TestOptions> {
   public constructor(options: TestOptions) {
     super(TestBackend, options);
     TestClient.instance = this;
-  }
-
-  public getSdkInfo(): SdkInfo {
-    return TEST_SDK;
   }
 }
