+++ packages/node/src/client.ts
@@ -1,8 +1,5 @@
 import { BaseClient } from '@sentry/core';
-import { SdkInfo } from '@sentry/types';
 import { NodeBackend, NodeOptions } from './backend';
-
-export const SDK_VERSION = '4.0.0-beta.9';
 
 /**
  * The Sentry Node SDK Client.
@@ -17,15 +14,5 @@ export class NodeClient extends BaseClient<NodeBackend, NodeOptions> {
    */
   public constructor(options: NodeOptions) {
     super(NodeBackend, options);
-  }
-
-  /**
-   * @inheritDoc
-   */
-  public getSdkInfo(): SdkInfo {
-    return {
-      name: 'sentry.javascript.node',
-      version: SDK_VERSION,
-    };
   }
 }
