+++ packages/node/src/integrations/sdkpackage.ts
@@ -1,6 +1,6 @@
 import { getDefaultHub } from '@sentry/hub';
 import { Integration, SentryEvent } from '@sentry/types';
-import { SDK_VERSION } from '../client';
+import { SDK_NAME, SDK_VERSION } from '../version';
 
 /** Adds SDK package info to an event. */
 export class SDKPackage implements Integration {
@@ -14,15 +14,17 @@ export class SDKPackage implements Integration {
    */
   public install(): void {
     getDefaultHub().addEventProcessor(async (event: SentryEvent) => {
-      if (event.sdk) {
-        event.sdk.packages = [
-          ...(event.sdk.packages || []),
+      event.sdk = {
+        name: SDK_NAME,
+        packages: [
+          ...((event.sdk && event.sdk.packages) || []),
           {
             name: 'npm:@sentry/node',
             version: SDK_VERSION,
           },
-        ];
-      }
+        ],
+        version: SDK_VERSION,
+      };
       return event;
     });
   }
