+++ packages/node/src/transports/base.ts
@@ -9,7 +9,7 @@ import {
 import { serialize } from '@sentry/utils/object';
 import * as http from 'http';
 import * as https from 'https';
-import { getDefaultHub, NodeClient } from '../index';
+import { SDK_NAME, SDK_VERSION } from '../version';
 
 /** Internal used interface for typescript */
 export interface HTTPRequest {
@@ -36,14 +36,9 @@ export abstract class BaseTransport implements Transport {
   private getAuthHeader(): string {
     const header = ['Sentry sentry_version=7'];
     header.push(`sentry_timestamp=${new Date().getTime()}`);
-    const client = getDefaultHub().getClient() as NodeClient;
-    if (client) {
-      header.push(
-        `sentry_client=${client.getSdkInfo().name}/${
-          client.getSdkInfo().version
-        }`,
-      );
-    }
+
+    header.push(`sentry_client=${SDK_NAME}/${SDK_VERSION}`);
+
     header.push(`sentry_key=${this.dsn.user}`);
     if (this.dsn.pass) {
       header.push(`sentry_secret=${this.dsn.pass}`);
