+++ packages/node/src/version.ts
@@ -1,0 +1,2 @@
+export const SDK_NAME = 'sentry.javascript.node';
+export const SDK_VERSION = '4.0.0-beta.9';
